* Refactored gathering attributes to another trait * Using the trait to set and get attributes in sqs and statistics enpoint
