+++ rest/rest-sqs/src/main/scala/org/elasticmq/rest/sqs/StatisticsDirectives.scala
@@ -3,14 +3,15 @@ package org.elasticmq.rest.sqs
 import java.util.concurrent.Executors
 
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
-import akka.http.scaladsl.model.StatusCodes.{InternalServerError, NotFound}
+import akka.http.scaladsl.model.StatusCodes.NotFound
+import akka.http.scaladsl.server.Route
 import org.elasticmq.QueueStatistics
 import org.elasticmq.metrics.QueueMetricsOps
 import org.elasticmq.rest.sqs.directives.ElasticMQDirectives
 import org.elasticmq.util.NowProvider
 import spray.json._
 
-import scala.concurrent.{Await, ExecutionContext}
+import scala.concurrent.ExecutionContext
 import scala.util.{Failure, Success}
 
 final case class QueueResponse(
@@ -31,7 +32,7 @@ trait JsonSupport extends SprayJsonSupport with DefaultJsonProtocol {
 }
 
 trait StatisticsDirectives extends JsonSupport {
-  this: ElasticMQDirectives =>
+  this: ElasticMQDirectives with QueueAttributesOps =>
 
   lazy val nowProvider = new NowProvider
   lazy val ec: ExecutionContext = ExecutionContext.fromExecutorService(Executors.newFixedThreadPool(2))
@@ -46,10 +47,7 @@ trait StatisticsDirectives extends JsonSupport {
           }
         },
         path(Segment) { queueName =>
-          onComplete(gatherSpecificQueueWithStats(queueName)) {
-            case Success(value) => complete(value)
-            case Failure(ex) => complete(NotFound, s"Can't load data for queue ${queueName}. Error ${ex.getMessage}")
-          }
+          gatherSpecificQueueWithStats(queueName)
         }
       )
     }
@@ -61,12 +59,16 @@ trait StatisticsDirectives extends JsonSupport {
       .map { x => x.map { case (name, stats) => mapToRest(name, Some(stats)) } }
   }
 
-  def gatherSpecificQueueWithStats(queueName: String) = {
-    //TODO gather attribs
-
-    QueueMetricsOps.getQueueStatistics(queueName, queueManagerActor, nowProvider)
-      .map(_ => mapToRest(queueName, None))
-
+  def gatherSpecificQueueWithStats(queueName: String): Route = {
+    val map = Map("QueueName" -> queueName)
+    queueActorAndDataFromRequest(map) { (queueActor, queueData) =>
+      onComplete(getAllQueueAttributes(map, queueActor, queueData)) {
+        case Success(value) => complete(value)
+        case Failure(ex) =>
+          logger.error(s"Error while loading statistics for queue ${queueName}", ex)
+          complete(NotFound, s"Can't load data for queue ${queueName}")
+      }
+    }
   }
 
   private def mapToRest(queueName: String, maybeStatistics: Option[QueueStatistics]) = {
